Avoid panic() on some API call failures
Due to an ordering race, we can skip waiting for the config call to
finish before returning from Collect(), so the registry can close the
metrics channel before the goroutine that returns the error is able to
return. This should ensure that we always wait, even if we end up
waiting twice as a result.

Fixes: https://github.com/frebib/nzbget-exporter/issues/5
Signed-off-by: Joe Groocock <me@frebib.net>